Reorganizing src/utils, by renaming submodules
And moving formatting stuff out of fs.rs
