+++ src/utils/mod.rs
@@ -1,9 +1,11 @@
-//! Random filesystem-related stuff used on ouch.
+//! Random and miscellaneous utils used in ouch.
 
-mod bytes;
+mod formatting;
 mod fs;
-mod question_policy;
+mod question;
 
-pub use bytes::Bytes;
-pub use fs::*;
-pub use question_policy::*;
+pub use formatting::{concatenate_os_str_list, nice_directory_display, strip_cur_dir, to_utf, Bytes};
+pub use fs::{cd_into_same_dir_as, colors, create_dir_if_non_existent, dir_is_empty, try_infer_extension};
+pub use question::{
+    create_or_ask_overwrite, user_wants_to_continue_decompressing, user_wants_to_overwrite, QuestionPolicy,
+};
